Parsing of gga and zda works, vtg still problem
by changing the scanf functions a little we managed to gat the parsing working for both GGA and ZDA messages. 
VTG is still a problem because arduisimple doesn't fill in all the parameters.
received VTG messag : $GPVTG,,T,,M,0.017,N,0.031,K,D*22

Also changed loop so it checks on what message it is before trying to parse it.